Switch to calling GetUint64() then casting to avoid debug assertions

+++ src/internal/sio_packet.cpp
@@ -143,7 +143,7 @@ namespace sio
         }
         else if (value.IsUint64())
         {
-          return int_message::create(value.GetInt64());
+          return int_message::create(static_cast<int64_t>(value.GetUint64()));
         }
         else if(value.IsDouble())
         {
@@ -187,14 +187,14 @@ namespace sio
             }
             return ptr;
         }
-		else if(value.IsBool())
-		{
-			return bool_message::create(value.GetBool());
-		}
-		else if(value.IsNull())
-		{
-			return null_message::create();
-		}
+        else if(value.IsBool())
+        {
+            return bool_message::create(value.GetBool());
+        }
+        else if(value.IsNull())
+        {
+            return null_message::create();
+        }
         return message::ptr();
     }
